// src/app/Core/Microservices/Restaurant/Models/Client/client.request.ts
export class ClientRequestPost {
  document: string = "";
  typeDocument: string = "";
  name: string = "";
  lastName: string = "";
  contact: string = "";
  address: string = "";
  user: string = "";
}

export class ClientRequestPut {
  document: string = "";
  typeDocument: string = "";
  name: string = "";
  lastName: string = "";
  contact: string = "";
  address: string = "";
  user: string = "";
}